requirements.txt
"""conda-forge"
geopandas
pandas
django
psycopg2
django-extensions
django-cors-headers
djangorestframework
djangorestframework-gis
djangorestframework_simplejwt
